Added ECU-Panel Added Animation to ECU-Panel Added information to display to ECU-Panel Added pages logic to display different values changeable with arrows Added TouchScreen logic Added control for modules. Added Faults to Panel displaying wear status Added Faults to panel displaying conditions of some parts (more to come) Added logic for disabling the airFuel calculation -> will make A/F stay fixed without fluctuation thanks to @Rodrigz for this suggestion Added Mod communication for values of turbocharger-mod Added ECU-Panel page for turbocharger-mod Added stage2Rev rpm to ecu-panel. This value can be adjusted by pressing on it and then using + - buttons Added screws to info panel
Changed cruise control throttle calculation to be more simple. This also fixes the problem with "car not reaching set speed"
Changed set cruise control logic. Now everytime if you are driving above 20km/h the cruise control speed will be set to the current speed driven.
Changed debug information position
Changed ALS logic to mimic Antilag on Deceleration instead of by using a "button"
Changed ProductImages of all parts
Changed max rpm to 8500 as 9000 would blow up the engine most of the time
Changed stage2 to 2step
Changed ALS logic to only have an affect if the turbocharger mod is installed (to create flames and backfire). This is now 100% for showoff and does not have an effect on driving

Fixed Cruise Control not applying throttle (this is a workaround as the previous way for some reason no longer works. This might cause an unpleasent throttle sound now)
Fixed Mod Communication between ecu and turbo
Fixed ALS not disabling for Turbo when control panel was removed
Fixed maxRPM not beeing increased when smart engine module is installed

Removed the original control panel including all of the code for it
Updated to ScrewablePartAPI v1.3
ALS is now more for showoff
Complete Code rework
Improved performance ALOT